Crate sp_timestamp

Substrate core types and inherents for timestamps.

Structs

InherentDataProvider

Provide duration since unix epoch in millisecond for timestamp inherent.

Timestamp

Unit type wrapper that represents a timestamp.

Enums

InherentError

Errors that can occur while checking the timestamp inherent.

Constants

INHERENT_IDENTIFIER

The identifier for the timestamp inherent.

Traits

TimestampInherentData

Auxiliary trait to extract timestamp inherent data.

Type Aliases

InherentType

The type of the inherent.