Crate sp_timestamp

Substrate core types and inherents for timestamps.

Structs

InherentDataProvider
Provide duration since unix epoch in millisecond for timestamp inherent.

Enums

InherentError
Errors that can occur while checking the timestamp inherent.

Constants

INHERENT_IDENTIFIER
The identifier for the timestamp inherent.

Traits

OnTimestampSet
A trait which is called when the timestamp is set.

TimestampInherentData
Auxiliary trait to extract timestamp inherent data.

Type Definitions

InherentType
The type of the inherent.